// src/interfaces/PromoFinal.interface.tsx
import {Subcategory} from './Subcategory.interface';
// Generated by https://quicktype.io

export interface PromoFinalResponse {
  count: number;
  page: number;
  totalPages: number;
  data: PromoFinal[];
}

export interface PromoFinal {
  status: boolean;
  createdAt: string;
  updatedAt: string;
  image: Image;
  subcategory: Subcategory;
  id: string;
}

export interface Image {
  url: string;
  id: string;
}